+++ capi/example/examples/soft_stm.h
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 07/09/2022
+// Last Modified: 24/10/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -12,10 +12,6 @@
 #pragma once
 
 #include <math.h>
-
-#define TRANS_SPACING_MM (10.16)
-#define NUM_TRANS_X (18)
-#define NUM_TRANS_Y (14)
 
 void* soft_stm(void* autd) {
   void* s = NULL;
@@ -28,8 +24,8 @@ void* soft_stm(void* autd) {
   AUTDSend(autd, m, NULL);
   AUTDDeleteModulation(m);
 
-  double x = TRANS_SPACING_MM * (((double)NUM_TRANS_X - 1.0) / 2.0);
-  double y = TRANS_SPACING_MM * (((double)NUM_TRANS_Y - 1.0) / 2.0);
+  double x = 90.0;
+  double y = 70.0;
   double z = 150.0;
 
   void* stm = NULL;