+++ cs/src/FirmwareInfo.cs
@@ -4,7 +4,7 @@
  * Created Date: 28/04/2021
  * Author: Shun Suzuki
  * -----
- * Last Modified: 02/06/2022
+ * Last Modified: 25/10/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2021 Shun Suzuki. All rights reserved.
@@ -12,6 +12,10 @@
  */
 
 using System;
+
+#if UNITY_2020_2_OR_NEWER
+#nullable enable
+#endif
 
 namespace AUTD3Sharp
 {
@@ -32,3 +36,7 @@ namespace AUTD3Sharp
         public override int GetHashCode() => Info.GetHashCode();
     }
 }
+
+#if UNITY_2020_2_OR_NEWER
+#nullable disable
+#endif