+++ cs/src/NativeMethods/Base.cs
@@ -1,12 +1,16 @@
 // This file was automatically generated from header file
 using System;
 using System.Runtime.InteropServices;
-            
+
+#if UNITY_2020_2_OR_NEWER
+#nullable enable
+#endif
+
 namespace AUTD3Sharp.NativeMethods
 {
     internal static class Base
     {
-        const string DLL = "autd3capi";
+        private const string DLL = "autd3capi";
 
         [DllImport(DLL, CharSet = CharSet.Ansi, BestFitMapping = false, ThrowOnUnmappableChar = true, CallingConvention = CallingConvention.Cdecl)] public static extern int AUTDGetLastError(System.Text.StringBuilder? error);
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] public static extern void AUTDCreateController(out IntPtr @out);
@@ -102,3 +106,7 @@ namespace AUTD3Sharp.NativeMethods
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] public static extern void AUTDDeleteSoftwareSTM(IntPtr stm);
     }
 }
+
+#if UNITY_2020_2_OR_NEWER
+#nullable disable
+#endif