+++ matlab/examples/bessel.m
@@ -4,7 +4,7 @@
 %Created Date: 11/06/2022
 %Author: Shun Suzuki
 %-----
-%Last Modified: 11/06/2022
+%Last Modified: 24/10/2022
 %Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 %-----
 %Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -16,12 +16,8 @@ function bessel(cnt)
     cnt.send(config);
     config.delete();
 
-    TRANS_SPACING_MM = 10.16;
-    NUM_TRANS_X = 18;
-    NUM_TRANS_Y = 14;
-    x = TRANS_SPACING_MM * ((NUM_TRANS_X - 1.0) / 2.0);
-    y = TRANS_SPACING_MM * ((NUM_TRANS_Y - 1.0) / 2.0);
-
+    x = 90.0;
+    y = 70.0;
     g = Bessel([x y 0], [0 0 1], 13.0/180.0 * pi);
     m = Sine(150);