+++ matlab/examples/holo.m
@@ -4,7 +4,7 @@
 %Created Date: 11/06/2022
 %Author: Shun Suzuki
 %-----
-%Last Modified: 22/06/2022
+%Last Modified: 24/10/2022
 %Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 %-----
 %Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -41,11 +41,8 @@ function holo(cnt)
     f = str2func(sprintf('@(x)%s', opts(i, 1)));
     g = f(backend);
 
-    TRANS_SPACING_MM = 10.16;
-    NUM_TRANS_X = 18;
-    NUM_TRANS_Y = 14;
-    x = TRANS_SPACING_MM * ((NUM_TRANS_X - 1.0) / 2.0);
-    y = TRANS_SPACING_MM * ((NUM_TRANS_Y - 1.0) / 2.0);
+    x = 90.0;
+    y = 70.0;
     z = 150.0;
 
     center = [x y z];