+++ python/pyautd3/__init__.py
@@ -4,7 +4,7 @@ Project: pyautd3
 Created Date: 24/05/2021
 Author: Shun Suzuki
 -----
-Last Modified: 21/10/2022
+Last Modified: 25/10/2022
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -15,7 +15,7 @@ Copyright (c) 2022 Shun Suzuki. All rights reserved.
 from pyautd3.autd import SilencerConfig
 from pyautd3.autd import Controller
 from pyautd3.autd import Amplitudes, ModDelayConfig
-from pyautd3.autd import NUM_TRANS_IN_UNIT, NUM_TRANS_X, NUM_TRANS_Y, TRANS_SPACING_MM, DEVICE_HEIGHT, DEVICE_WIDTH
+from pyautd3.autd import NUM_TRANS_IN_UNIT, NUM_TRANS_X, NUM_TRANS_Y, TRANS_SPACING, DEVICE_HEIGHT, DEVICE_WIDTH
 
 __all__ = [
     'SilencerConfig',
@@ -25,9 +25,9 @@ __all__ = [
     'NUM_TRANS_IN_UNIT',
     'NUM_TRANS_X',
     'NUM_TRANS_Y',
-    'TRANS_SPACING_MM',
+    'TRANS_SPACING',
     'DEVICE_WIDTH',
     'DEVICE_HEIGHT'
 ]
 
-__version__ = '2.4.4'
+__version__ = '2.4.5'