+++ python/pyautd3/gain/holo/__init__.py
@@ -4,7 +4,7 @@ Project: holo
 Created Date: 21/10/2022
 Author: Shun Suzuki
 -----
-Last Modified: 21/10/2022
+Last Modified: 25/10/2022
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -14,7 +14,7 @@ Copyright (c) 2022 Shun Suzuki. All rights reserved.
 from .backend import EigenBackend
 from .cuda_backend import CUDABackend
 
-from .holo import Clamp, DontCare, Normalize, Uniform
+from .constraint import Clamp, DontCare, Normalize, Uniform
 
 from .apo import APO
 from .evd import EVD
@@ -25,7 +25,6 @@ from .lm import LM
 from .greedy import Greedy
 from .lss_greedy import LSSGreedy
 from .naive import Naive
-
 
 __all__ = [
     'EigenBackend',