+++ src/extra/simulator/vulkan_imgui.hpp
@@ -3,7 +3,7 @@
 // Created Date: 03/10/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 18/10/2022
+// Last Modified: 25/10/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -167,32 +167,37 @@ class VulkanImGui {
     ImGui_ImplVulkan_DestroyFontUploadObjects();
   }
 
-  void load_settings(const SimulatorSettings& setting) {
-    slice_pos = glm::vec3(setting.slice_pos_x, setting.slice_pos_y, setting.slice_pos_z);
-    slice_rot = glm::vec3(setting.slice_rot_x, setting.slice_rot_y, setting.slice_rot_z);
-    slice_width = setting.slice_width;
-    slice_height = setting.slice_height;
-    pixel_size = setting.slice_pixel_size;
-    color_scale = setting.slice_color_scale;
-    slice_alpha = setting.slice_alpha;
-    show_radiation_pressure = setting.show_radiation_pressure;
-    coloring_method = setting.coloring_method;
+  void load_settings(const SimulatorSettings& settings) {
+    slice_pos = glm::vec3(settings.slice_pos_x, settings.slice_pos_y, settings.slice_pos_z);
+    slice_rot = glm::vec3(settings.slice_rot_x, settings.slice_rot_y, settings.slice_rot_z);
+    slice_width = settings.slice_width;
+    slice_height = settings.slice_height;
+    pixel_size = settings.slice_pixel_size;
+    color_scale = settings.slice_color_scale;
+    slice_alpha = settings.slice_alpha;
+    show_radiation_pressure = settings.show_radiation_pressure;
+    coloring_method = settings.coloring_method;
     coloring_method_idx = convert(coloring_method);
 
-    camera_pos = glm::vec3(setting.camera_pos_x, setting.camera_pos_y, setting.camera_pos_z);
-    camera_rot = glm::vec3(setting.camera_rot_x, setting.camera_rot_y, setting.camera_rot_z);
-    fov = setting.camera_fov;
-    near_clip = setting.camera_near_clip;
-    far_clip = setting.camera_far_clip;
-    _cam_move_speed = setting.camera_move_speed;
+    camera_pos = glm::vec3(settings.camera_pos_x, settings.camera_pos_y, settings.camera_pos_z);
+    camera_rot = glm::vec3(settings.camera_rot_x, settings.camera_rot_y, settings.camera_rot_z);
+    fov = settings.camera_fov;
+    near_clip = settings.camera_near_clip;
+    far_clip = settings.camera_far_clip;
+    _cam_move_speed = settings.camera_move_speed;
 
-    _font_size = setting.font_size;
-    background = glm::vec4(setting.background_r, setting.background_g, setting.background_b, setting.background_a);
+    sound_speed = settings.sound_speed;
 
-    _show_mod_plot = setting.show_mod_plot;
-    _show_mod_plot_raw = setting.show_mod_plot_raw;
+    _font_size = settings.font_size;
+    background = glm::vec4(settings.background_r, settings.background_g, settings.background_b, settings.background_a);
 
-    setting.image_save_path.copy(save_path, 256);
+    _show_mod_plot = settings.show_mod_plot;
+    _show_mod_plot_raw = settings.show_mod_plot_raw;
+
+    use_meter = settings.use_meter;
+    use_left_handed = settings.use_left_handed;
+
+    settings.image_save_path.copy(save_path, 256);
   }
 
   void save_settings(SimulatorSettings& settings) const {
@@ -221,6 +226,7 @@ class VulkanImGui {
     settings.camera_far_clip = far_clip;
     settings.camera_move_speed = _cam_move_speed;
 
+    settings.sound_speed = sound_speed;
     settings.font_size = _font_size;
 
     settings.background_r = background.r;
@@ -232,11 +238,14 @@ class VulkanImGui {
     settings.show_mod_plot_raw = _show_mod_plot_raw;
 
     settings.image_save_path = std::string(save_path);
+
+    settings.use_meter = use_meter;
+    settings.use_left_handed = use_left_handed;
   }
 
   void init(const uint32_t image_count, const VkRenderPass renderer_pass, const SimulatorSettings& settings) {
     load_settings(settings);
-    _default_settings = settings;
+    _initial_settings = settings;
 
     IMGUI_CHECKVERSION();
     ImGui::CreateContext();
@@ -307,7 +316,7 @@ class VulkanImGui {
 
       if (!io.WantCaptureMouse) {
         const auto mouse_wheel = io.MouseWheel;
-        const auto trans = -f * mouse_wheel * _cam_move_speed;
+        const auto trans = use_left_handed ? f * mouse_wheel * _cam_move_speed : -f * mouse_wheel * _cam_move_speed;
         camera_pos[0] += trans.x;
         camera_pos[1] += trans.y;
         camera_pos[2] += trans.z;
@@ -318,7 +327,7 @@ class VulkanImGui {
         const auto mouse_delta = io.MouseDelta;
         if (io.MouseDown[0]) {
           if (io.KeyShift) {
-            const auto delta = glm::vec2(mouse_delta.x, mouse_delta.y) * _cam_move_speed / 3000.0f;
+            const auto delta = glm::vec2(mouse_delta.x, mouse_delta.y) * _cam_move_speed / scale() / 3000.0f;
             const auto to = -r * delta.x + u * delta.y + f;
             const auto rotation = helper::quaternion_to(f, to);
             camera_rot = degrees(eulerAngles(rotation * rot));
@@ -340,9 +349,9 @@ class VulkanImGui {
     if (ImGui::BeginTabBar("Settings")) {
       if (ImGui::BeginTabItem("Slice")) {
         ImGui::Text("Position");
-        if (ImGui::DragFloat("X##Slice", &slice_pos.x)) flag.set(UpdateFlags::UPDATE_SLICE_POS);
-        if (ImGui::DragFloat("Y##Slice", &slice_pos.y)) flag.set(UpdateFlags::UPDATE_SLICE_POS);
-        if (ImGui::DragFloat("Z##Slice", &slice_pos.z)) flag.set(UpdateFlags::UPDATE_SLICE_POS);
+        if (ImGui::DragFloat("X##Slice", &slice_pos.x, 1 * scale())) flag.set(UpdateFlags::UPDATE_SLICE_POS);
+        if (ImGui::DragFloat("Y##Slice", &slice_pos.y, 1 * scale())) flag.set(UpdateFlags::UPDATE_SLICE_POS);
+        if (ImGui::DragFloat("Z##Slice", &slice_pos.z, 1 * scale())) flag.set(UpdateFlags::UPDATE_SLICE_POS);
         ImGui::Separator();
 
         ImGui::Text("Rotation");
@@ -352,10 +361,16 @@ class VulkanImGui {
         ImGui::Separator();
 
         ImGui::Text("Size");
-        if (ImGui::DragInt("Width##Slice", &slice_width, 1, 1, 2000)) flag.set(UpdateFlags::UPDATE_SLICE_SIZE);
-        if (ImGui::DragInt("Height##Slice", &slice_height, 1, 1, 2000)) flag.set(UpdateFlags::UPDATE_SLICE_SIZE);
-        if (ImGui::DragFloat("Pixel size##Slice", &pixel_size, 1, 0.1f, 8)) {
-          if (pixel_size <= 0.1f) pixel_size = 0.1f;
+        if (ImGui::DragFloat("Width##Slice", &slice_width, 1 * scale(), 1 * scale(), 2000 * scale())) {
+          if (slice_width < 1 * scale()) slice_width = 1 * scale();
+          flag.set(UpdateFlags::UPDATE_SLICE_SIZE);
+        }
+        if (ImGui::DragFloat("Height##Slice", &slice_height, 1 * scale(), 1 * scale(), 2000 * scale())) {
+          if (slice_height < 1 * scale()) slice_height = 1 * scale();
+          flag.set(UpdateFlags::UPDATE_SLICE_SIZE);
+        }
+        if (ImGui::DragFloat("Pixel size##Slice", &pixel_size, 1 * scale(), 0.1f * scale(), 8 * scale())) {
+          if (pixel_size <= 0.1f * scale()) pixel_size = 0.1f * scale();
           flag.set(UpdateFlags::UPDATE_SLICE_SIZE);
         }
         ImGui::Separator();
@@ -407,9 +422,9 @@ class VulkanImGui {
 
       if (ImGui::BeginTabItem("Camera")) {
         ImGui::Text("Position");
-        if (ImGui::DragFloat("X##Camera", &camera_pos.x)) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
-        if (ImGui::DragFloat("Y##Camera", &camera_pos.y)) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
-        if (ImGui::DragFloat("Z##Camera", &camera_pos.z)) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
+        if (ImGui::DragFloat("X##Camera", &camera_pos.x, 1 * scale())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
+        if (ImGui::DragFloat("Y##Camera", &camera_pos.y, 1 * scale())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
+        if (ImGui::DragFloat("Z##Camera", &camera_pos.z, 1 * scale())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
         ImGui::Separator();
         ImGui::Text("Rotation");
         if (ImGui::DragFloat("RX##Camera", &camera_rot.x, 1, -180, 180)) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
@@ -418,21 +433,22 @@ class VulkanImGui {
         ImGui::Separator();
         ImGui::Text("Perspective");
         if (ImGui::DragFloat("FOV", &fov, 1, 0, 180)) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
-        if (ImGui::DragFloat("Near clip", &near_clip, 1, 0, std::numeric_limits<float>::infinity())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
-        if (ImGui::DragFloat("Far clip", &far_clip, 1, 0, std::numeric_limits<float>::infinity())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
+        if (ImGui::DragFloat("Near clip", &near_clip, 1 * scale(), 0, std::numeric_limits<float>::infinity()))
+          flag.set(UpdateFlags::UPDATE_CAMERA_POS);
+        if (ImGui::DragFloat("Far clip", &far_clip, 1 * scale(), 0, std::numeric_limits<float>::infinity())) flag.set(UpdateFlags::UPDATE_CAMERA_POS);
         flag.set(UpdateFlags::UPDATE_CAMERA_POS);
         ImGui::Separator();
-        ImGui::DragFloat("Move speed", &_cam_move_speed);
+        ImGui::DragFloat("Move speed", &_cam_move_speed, 1 * scale());
         ImGui::EndTabItem();
       }
 
       if (ImGui::BeginTabItem("Config")) {
-        if (ImGui::DragFloat("Sound speed", &sound_speed, 1)) {
+        if (ImGui::DragFloat("Sound speed", &sound_speed, 1 * scale())) {
           for (size_t dev = 0; dev < cpus.size(); dev++) {
             const auto& cycles = cpus[dev].fpga().cycles();
             for (size_t i = 0; i < driver::NUM_TRANS_IN_UNIT; i++) {
               const auto freq = static_cast<float>(driver::FPGA_CLK_FREQ) / static_cast<float>(cycles[i]);
-              sources.drives()[i + dev * driver::NUM_TRANS_IN_UNIT].set_wave_num(freq, sound_speed * 1e3f);
+              sources.drives()[i + dev * driver::NUM_TRANS_IN_UNIT].set_wave_num(freq, sound_speed);
             }
           }
           flag.set(UpdateFlags::UPDATE_SOURCE_DRIVE);
@@ -528,7 +544,10 @@ class VulkanImGui {
             ImGui::Text("Gain STM");
           else {
             ImGui::Text("Point STM");
-            ImGui::Text("Sound speed: %.3lf [mm/s]", cpus[0].fpga().sound_speed() * 1000.0 / 1024.0);
+            if (use_meter)
+              ImGui::Text("Sound speed: %.3lf [m/s]", cpus[0].fpga().sound_speed() / 1024.0);
+            else
+              ImGui::Text("Sound speed: %.3lf [mm/s]", cpus[0].fpga().sound_speed() * 1000.0 / 1024.0);
           }
 
           ImGui::Text("Size: %d", static_cast<int32_t>(cpus[0].fpga().stm_cycle()));
@@ -606,15 +625,22 @@ class VulkanImGui {
     if (ImGui::SmallButton("Auto")) {
       const auto sr = mat4_cast(glm::quat(radians(slice_rot)));
       const auto srf = glm::vec3(sr * glm::vec4(0, 0, 1, 1));
-      camera_pos = slice_pos + srf * 600.0f;
+      camera_pos = slice_pos + srf * 600.0f * scale();
       camera_rot = slice_rot;
-
       flag.set(UpdateFlags::UPDATE_CAMERA_POS);
     }
     ImGui::SameLine();
     if (ImGui::SmallButton("Reset")) {
-      load_settings(_default_settings);
+      load_settings(_initial_settings);
       flag.set(UpdateFlags::all().value());
+      flag.remove(UpdateFlags::SAVE_IMAGE);
+    }
+    ImGui::SameLine();
+    if (ImGui::SmallButton("Default")) {
+      _initial_settings.load_default(use_meter, use_left_handed);
+      load_settings(_initial_settings);
+      flag.set(UpdateFlags::all().value());
+      flag.remove(UpdateFlags::SAVE_IMAGE);
     }
 
     ImGui::End();
@@ -637,8 +663,8 @@ class VulkanImGui {
   }
 
   [[nodiscard]] std::pair<glm::mat4, glm::mat4> get_view_proj(const float aspect) const {
-    const auto rot = glm::quat(radians(camera_rot));
-    const auto p = camera_pos;
+    const auto rot = to_gl_rot(glm::quat(radians(camera_rot)));
+    const auto p = to_gl_pos(camera_pos);
     const auto view = helper::orthogonal(p, rot);
     auto proj = glm::perspective(glm::radians(fov), aspect, near_clip, far_clip);
     proj[1][1] *= -1;
@@ -646,17 +672,25 @@ class VulkanImGui {
   }
 
   [[nodiscard]] glm::mat4 get_slice_model() const {
-    return translate(glm::identity<glm::mat4>(), slice_pos) * mat4_cast(glm::quat(radians(slice_rot)));
+    return translate(glm::identity<glm::mat4>(), to_gl_pos(slice_pos)) * mat4_cast(to_gl_rot(glm::quat(radians(slice_rot))));
   }
 
-  int32_t slice_width{0};
-  int32_t slice_height{0};
+  [[nodiscard]] float scale() const noexcept { return use_meter ? 1e-3f : 1.0f; }
+
+  [[nodiscard]] glm::vec3 to_gl_pos(const glm::vec3 v) const { return use_left_handed ? glm::vec3(v.x, v.y, -v.z) : v; }
+  [[nodiscard]] glm::quat to_gl_rot(const glm::quat rot) const { return use_left_handed ? glm::quat(rot.w, -rot.x, -rot.y, rot.z) : rot; }
+
+  float slice_width{0};
+  float slice_height{0};
   glm::vec3 slice_pos{};
   glm::vec3 slice_rot{};
   float color_scale{2.0};
   float slice_alpha{1.0f};
   float pixel_size{1.0};
   bool show_radiation_pressure{false};
+
+  bool use_meter{false};
+  bool use_left_handed{false};
 
   glm::vec3 camera_pos{};
   glm::vec3 camera_rot{};
@@ -683,7 +717,7 @@ class VulkanImGui {
  private:
   const helper::WindowHandler* _window;
   const helper::VulkanContext* _context;
-  SimulatorSettings _default_settings;
+  SimulatorSettings _initial_settings;
   float _font_size = 16.0f;
   bool _update_font = false;
   float _cam_move_speed = 10.0f;